Broke out the generateDemoBoard function from main.js. #28
